backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
playwright==1.44.0
anthropic==0.28.0
pymupdf==1.24.5
python-dotenv==1.0.1
pydantic==2.7.3
aiofiles==23.2.1
httpx==0.27.0
